requirements.txt
absl-py==1.4.0
altair==5.0.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
jax==0.4.11
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
scipy==1.10.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.23.1
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
